edit coding part

+++ CarsManagment/Cars.cs
@@ -16,47 +16,21 @@ namespace CarsManagment
     {
         MySqlConnection dbconnection;
         DataGridViewRow row1 = null;
-        Car_Record carRecord = null;
-        Car_Update carUpdate = null;
+        public Car_Record carRecord = null;
+        public Car_Update carUpdate = null;
 
-        CarDrivers carDrivers = null;
-        CarLicenseRecord carLicenseRecord = null;
-        CarSpareParts carSpareParts = null;
+        public CarDrivers carDrivers = null;
+        public Car_Papers carPaper = null;
+        public CarLicenseRecord carLicenseRecord = null;
+        public CarLicenseUpdate carLicenseUpdate = null;
+        public CarSpareParts carSpareParts = null;
+
         public Cars()
         {
-            InitializeComponent();
-            dbconnection = new MySqlConnection(connection.connectionString);
-        }
-
-        private void Cars_Load(object sender, EventArgs e)
-        {
             try
             {
-                dbconnection.Open();
-                DisplayCars();
-            }
-            catch (Exception ex)
-            {
-                MessageBox.Show(ex.ToString());
-            }
-            dbconnection.Close();
-        }
-
-        private void btnAdd_Click(object sender, EventArgs e)
-        {
-            try
-            {
-                if (carRecord == null)
-                {
-                    carRecord = new Car_Record();
-                    carRecord.Show();
-                    carRecord.Focus();
-                }
-                else
-                {
-                    carRecord.Show();
-                    carRecord.Focus();
-                }
+                InitializeComponent();
+                dbconnection = new MySqlConnection(connection.connectionString);
             }
             catch (Exception ex)
             {
@@ -64,6 +38,48 @@ namespace CarsManagment
             }
         }
 
+        private void Cars_Load(object sender, EventArgs e)
+        {
+            try
+            {
+                dbconnection.Open();
+                DisplayCars(dbconnection);
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.ToString());
+            }
+            dbconnection.Close();
+        }
+        private void txtSearch_TextChanged(object sender, EventArgs e)
+        {
+            try
+            {
+                dbconnection.Open();
+                dataGridView1.Rows.Clear();
+                string qeury = "select Car_Number,Car_Capacity,meter_reading,Openning_Account,Car_Value,DepreciationPeriod,PremiumDepreciation,cars.Car_ID from cars where Car_Number like'" + txtSearch.Text + "%'";
+                MySqlCommand com = new MySqlCommand(qeury, dbconnection);
+                MySqlDataReader dr = com.ExecuteReader();
+                while (dr.Read())
+                {
+                    int n = dataGridView1.Rows.Add();
+                    dataGridView1.Rows[n].Cells["Car_Number"].Value = dr[0].ToString();
+                    dataGridView1.Rows[n].Cells["Car_Capacity"].Value = dr[1].ToString();
+                    dataGridView1.Rows[n].Cells["meter_reading"].Value = dr[2].ToString();
+                    dataGridView1.Rows[n].Cells["Openning_Account"].Value = dr[3].ToString();
+                    dataGridView1.Rows[n].Cells["Car_Value"].Value = dr[4].ToString();
+                    dataGridView1.Rows[n].Cells["DepreciationPeriod"].Value = dr[5].ToString();
+                    dataGridView1.Rows[n].Cells["PremiumDepreciation"].Value = dr[6].ToString();
+                    dataGridView1.Rows[n].Cells["car_id"].Value = dr[7].ToString();
+                }
+                dr.Close();
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message);
+            }
+            dbconnection.Close();
+        }
         private void dataGridView1_CellClick(object sender, DataGridViewCellEventArgs e)
         {
             try
@@ -77,6 +93,28 @@ namespace CarsManagment
             }
         }
 
+        //main buttons
+        private void btnAdd_Click(object sender, EventArgs e)
+        {
+            try
+            {
+                if (carRecord == null)
+                {
+                    carRecord = new Car_Record(this);
+                    carRecord.Show();
+                    carRecord.Focus();
+                }
+                else
+                {
+                    carRecord.Show();
+                    carRecord.Focus();
+                }
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message);
+            }
+        }      
         private void btnUpdate_Click(object sender, EventArgs e)
         {
             try
@@ -85,7 +123,7 @@ namespace CarsManagment
                 {
                     if (row1 != null)
                     {
-                        carUpdate = new Car_Update(row1);
+                        carUpdate = new Car_Update(row1,this);
                         carUpdate.Show();
                         carUpdate.Focus();
                     }
@@ -105,17 +143,23 @@ namespace CarsManagment
                 MessageBox.Show(ex.Message);
             }
         }
-
         private void btnDelete_Click(object sender, EventArgs e)
         {
             try
             {
                 if (row1 != null)
                 {
-                    dbconnection.Open();
-                    string query = "delete from cars where Car_Number=" + row1.Cells[0].Value;
-                    MySqlCommand com = new MySqlCommand(query, dbconnection);
-                    com.ExecuteNonQuery();
+                    DialogResult dialogResult = MessageBox.Show("Are you sure you want to delete the item?", "", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+                    if (dialogResult == DialogResult.Yes)
+                    {
+                        dbconnection.Open();
+                        string query = "delete from cars where Car_Number=" + row1.Cells[0].Value;
+                        MySqlCommand com = new MySqlCommand(query, dbconnection);
+                        com.ExecuteNonQuery();
+                        DisplayCars(dbconnection);
+                    }
+                    else if (dialogResult == DialogResult.No)
+                    { }
                 }
                 else
                 {
@@ -138,7 +182,7 @@ namespace CarsManagment
                 {
                     if (row1 != null)
                     {
-                        carDrivers = new CarDrivers(row1);
+                        carDrivers = new CarDrivers(row1,this);
                         carDrivers.Show();
                         carDrivers.Focus();
                     }
@@ -159,17 +203,17 @@ namespace CarsManagment
             }
 
         }
-        private void btnCarLicense_Click(object sender, EventArgs e)
+        private void btnCarPaper_Click(object sender, EventArgs e)
         {
             try
             {
-                if (carLicenseRecord == null)
+                if (carPaper == null)
                 {
                     if (row1 != null)
                     {
-                        carLicenseRecord = new CarLicenseRecord(row1);
-                        carLicenseRecord.Show();
-                        carLicenseRecord.Focus();
+                        carPaper = new Car_Papers(row1,this);
+                        carPaper.Show();
+                        carPaper.Focus();
                     }
                     else
                     {
@@ -178,14 +222,63 @@ namespace CarsManagment
                 }
                 else
                 {
-                    carLicenseRecord.Show();
-                    carLicenseRecord.Focus();
+                    carPaper.Show();
+                    carPaper.Focus();
                 }
             }
             catch (Exception ex)
             {
                 MessageBox.Show(ex.Message);
             }
+        }
+        private void btnCarLicense_Click(object sender, EventArgs e)
+        {
+            try
+            {
+                if (row1 != null)
+                {
+                    dbconnection.Open();
+                    String query = "select Car_License_ID from car_license where Car_ID=" + row1.Cells["car_id"].Value;
+                    MySqlCommand com = new MySqlCommand(query, dbconnection);
+                    if (com.ExecuteScalar() == null)
+                    {
+                        if (carLicenseRecord == null)
+                        {
+                            carLicenseRecord = new CarLicenseRecord(row1, this);
+                            carLicenseRecord.Show();
+                            carLicenseRecord.Focus();
+                        }
+                        else
+                        {
+                            carLicenseRecord.Show();
+                            carLicenseRecord.Focus();
+                        }
+                    }
+                    else
+                    {
+                        if (carLicenseUpdate == null)
+                        {
+                            carLicenseUpdate = new CarLicenseUpdate(row1, this);
+                            carLicenseUpdate.Show();
+                            carLicenseUpdate.Focus();
+                        }
+                        else
+                        {
+                            carLicenseUpdate.Show();
+                            carLicenseUpdate.Focus();
+                        }
+                    }
+                }
+                else
+                {
+                    MessageBox.Show("Select row");
+                }
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message);
+            }
+            dbconnection.Close();
         }
         private void btnCarSparePart_Click(object sender, EventArgs e)
         {
@@ -195,7 +288,7 @@ namespace CarsManagment
                 {
                     if (row1 != null)
                     {
-                        carSpareParts = new CarSpareParts(row1);
+                        carSpareParts = new CarSpareParts(row1,this);
                         carSpareParts.Show();
                         carSpareParts.Focus();
                     }
@@ -215,8 +308,9 @@ namespace CarsManagment
                 MessageBox.Show(ex.Message);
             }
         }
+        
         //functions
-        public void DisplayCars()
+        public void DisplayCars(MySqlConnection dbconnection)
         {
             dataGridView1.Rows.Clear();
             string qeury = "select Car_Number,Car_Capacity,meter_reading,Openning_Account,Car_Value,DepreciationPeriod,PremiumDepreciation,cars.Car_ID from cars ";
@@ -237,6 +331,6 @@ namespace CarsManagment
             dr.Close();
         }
 
-       
+      
     }
 }
